Small updates

+++ roach/crypto/rsa.py
@@ -2,7 +2,6 @@
 # This file is part of Roach - https://github.com/jbremer/roach.
 # See the file 'docs/LICENSE.txt' for copying permission.
 
-from past.builtins import long
 from builtins import int
 import io
 
@@ -115,7 +114,7 @@ class RSA(object):
 
     @staticmethod
     def export_key(n, e, d=None, p=None, q=None, crt=None):
-        wrap = lambda x: None if x is None else long(x)
+        wrap = lambda x: None if x is None else int(x)
         tup = wrap(n), wrap(e), wrap(d), wrap(p), wrap(q), wrap(crt)
         return RSA_.construct(tup).exportKey()
